exchanged names between models and schemas
